move `check_mime_type` to `check.rs`
